+++ CHANGELOG.md
@@ -7,8 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [UNRELEASED]
 
 ### Added
-- Added context `clear!` for deleting all context data
 - Added Ruby 4.0 support
+- Added context `clear!` for deleting all context data
+- Added `Task.attribute_schema` class method to access attributes
+- Added attributes `to_h` method
 
 ### Updated
 - **IMPORTANT**: Switched from `MIT` to `LGPLv3` license